+++ Program.cs
@@ -1,6 +1,7 @@
 ﻿using LabManager.Database;
 using Microsoft.Data.Sqlite;
 using LabManager.Repositories;
+using LabManager.Models;
 
 var databaseConfig = new DatabaseConfig();
 new DatabaseSetup(databaseConfig);
@@ -32,18 +33,10 @@ if (modelName == "Computer")
         int id = Convert.ToInt32(args[2]);
         string ram = args[3];
         string processor = args[4];
+        
+        var computer = new Computer(id, ram, processor);
+        computerRepository.Save(computer);  
 
-        var connection = new SqliteConnection("Data Source=database.db");
-        connection.Open();
-
-        var command = connection.CreateCommand();
-        command.CommandText = "INSERT INTO Computers VALUES($id, $ram, $processor);";
-        command.Parameters.AddWithValue("$id", id);
-        command.Parameters.AddWithValue("$ram", ram);
-        command.Parameters.AddWithValue("$processor", processor);
-        command.ExecuteNonQuery();
-
-        connection.Close();
     }
 }
