+++ Repositories/ComputerRepository.cs
@@ -34,5 +34,20 @@ class ComputerRepository // classe que fala com o banco de dados
         return computers;
     }
 
+    public void Save (Computer computer) // recebe os prametros de computer e salva
+    {
+        var connection = new SqliteConnection(databaseConfig.ConnectionString);
+        connection.Open();
+
+        var command = connection.CreateCommand();
+        command.CommandText = "INSERT INTO Computers VALUES($id, $ram, $processor);";
+        command.Parameters.AddWithValue("$id", computer.Id);
+        command.Parameters.AddWithValue("$ram", computer.Ram);
+        command.Parameters.AddWithValue("$processor", computer.Processor);
+        command.ExecuteNonQuery();
+
+        connection.Close();
+    }
+
 }
